pyproject.toml
[tool.poetry]
name = "simple-flask"
version = "0.1.0"
description = "Simple python Flask"
authors = ["Oltho, Thomas TRAN <17247405+Oltho@users.noreply.github.com>"]
license = "MIT"
repository = "https://github.com/Oltho/simple-flask-docker"
homepage = "https://github.com/Oltho/simple-flask-docker"

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.1.1"

[tool.poetry.dev-dependencies]
pytest-testinfra = "^6.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
addopts = "--verbose"
norecursedirs = [
    "dist",
    "build",
    ".tox",
    "venv",
    ".venv"
]
testpaths = ["tests"]
